Update to Handle Year Level file and export data to CSV's
Now handling the Year Level (KCY) data, dropping not in use year levels (As defined by "DO NOT USE" in Description) and exporting files to CSV's.

Just left the credentialed AD access and the upload functions.

Upload needs to be tested against live data to ensure there is no corruption on upload